App file paths, added dependancies & imports
