+++ package.json
@@ -7,6 +7,7 @@
     "start": "node server/server.js",
     "develop": "concurrently \"cd server && npm run watch\" \"cd client && npm start\"",
     "seed": "cd server && npm run seed",
+    "install": "cd server && npm i && cd ../client && npm i",
     "build": "cd client && npm run build"
   },
   "repository": {
@@ -21,6 +22,7 @@
   },
   "homepage": "https://github.com/jkouneski/Kibble-Krew-Review#readme",
   "dependencies": {
-    "concurrently": "^6.2.0"
+    "concurrently": "^6.2.0",
+    "graphql": "^15.5.1"
   }
 }
