+++ server/package.json
@@ -13,7 +13,7 @@
   "license": "ISC",
   "dependencies": {
     "apollo-server-express": "^2.19.0",
-    "brcypt": "^1.0.1",
+    "bcrypt": "^5.0.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "graphql": "^15.5.1",
